requirements.txt
torch
torchvision
datasets[vision]
pillow
